+++ src/bielu.Umbraco.Cdn.Core/Controllers/BieluCdnManagmentController.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using bielu.Umbraco.Cdn.Core.Extensions;
 using bielu.Umbraco.Cdn.Core.Services;
@@ -13,6 +14,7 @@ using Umbraco.Cms.Web.Common.Controllers;
 using Umbraco.Extensions;
 
 namespace bielu.Umbraco.Cdn.Core.Controllers;
+
 [Route("cdn/api/management/[action]")]
 public class BieluCdnManagmentController : UmbracoApiController
 {
@@ -20,32 +22,47 @@ public class BieluCdnManagmentController : UmbracoApiController
     private readonly IUmbracoContextFactory _contextFactory;
     private readonly ICdnAuditService _auditService;
 
-    public BieluCdnManagmentController(ICdnManager manager, IUmbracoContextFactory contextFactory, Services.ICdnAuditService auditService)
+    public BieluCdnManagmentController(ICdnManager manager, IUmbracoContextFactory contextFactory,
+        Services.ICdnAuditService auditService)
     {
         _manager = manager;
         _contextFactory = contextFactory;
         _auditService = auditService;
     }
+
     public async Task<IEnumerable<AuditRecord>> GetAuditHistory()
     {
         return await _auditService.GetAllRecords();
     }
-    public async Task<IEnumerable<Provider>> GetProviders()
+
+    public async Task<IEnumerable<Provider>> GetProviders(int id = -1)
     {
-        return await _manager.GetProviders();
+        return await _manager.GetProviders(id);
     }
 
-    public async Task<Status> RefreshForNode(Guid id)
+    public async Task<Status> RefreshForNode(int id, string providerId = null, string domain = null)
     {
         List<Status> statuses = new List<Status>();
         using (var contextReference = _contextFactory.EnsureUmbracoContext())
         {
             var content = contextReference.UmbracoContext.Content.GetById(id);
             var domains = contextReference.UmbracoContext.Domains.GetAssigned(content.Id);
-            var urls = GenerateUrls(content,domains);
+            if (!string.IsNullOrWhiteSpace(domain))
+            {
+                domains = domains.Where(x => x.Name.Contains(domain));
+            }
+
+            var urls = GenerateUrls(content, domains);
+            if (!string.IsNullOrEmpty(providerId))
+            {
+                var service = await _manager.GetService(providerId);
+                statuses.AddRange(await service.PurgePages(urls));
+                return statuses.Merge();
+            }
+
             foreach (var service in await _manager.GetServices())
             {
-                statuses.AddRange(   await service.PurgePages(urls));
+                statuses.AddRange(await service.PurgePages(urls));
             }
         }
 
@@ -57,7 +74,7 @@ public class BieluCdnManagmentController : UmbracoApiController
         var urls = new List<string>();
         foreach (var domain in domains)
         {
-            urls.Add($"{domain.Name}{content.Url()}");
+            urls.Add($"https://{domain.Name}{content.Url(mode: UrlMode.Relative)}");
         }
 
         return urls;
